feat: Refactor Slack client
Slack client was bloated and was handling both rendering and computation in the
same file.

Split the slack code to make it more readable and easier to add more functionality.
